rename dll -> FancyZonesModuleInterface (#11488)
