--- src/modules/fancyzones/dll/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Windows.CppWinRT" version="2.0.200729.8" targetFramework="native" />
-  <package id="Microsoft.Windows.ImplementationLibrary" version="1.0.200902.2" targetFramework="native" />
-</packages>
