--- src/modules/fancyzones/dll/pch.cpp
@@ -1,1 +1,0 @@
-#include "pch.h"
