--- src/modules/fancyzones/dll/pch.h
@@ -1,17 +1,0 @@
-#pragma once
-
-#define WIN32_LEAN_AND_MEAN // Exclude rarely-used stuff from Windows headers
-#include <Unknwn.h>
-#include <winrt/base.h>
-#include <winrt/Windows.Foundation.h>
-#include <winrt/Windows.Foundation.Collections.h>
-#include <ProjectTelemetry.h>
-#include <TraceLoggingActivity.h>
-#include <wil\common.h>
-#include <wil\result.h>
-
-
-namespace winrt
-{
-    using namespace ::winrt;
-}
